Move iterator to pattern/IteratorPattern folder
